Accidentally modified example6.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@5932 1c3fd406-9068-4342-8353-511fcdf43622

+++ tests/example6.c
@@ -3,7 +3,7 @@
 #include <stdio.h> 
 
 
-void foo(long b[5], int value) {
+void foo(int b[5], int value) {
   int result = b[2];
   b[2] = value;
   return;
